+++ src/controller_device_handle.h
@@ -55,7 +55,7 @@ device_handle device_handle_new(clixon_handle h, const char *name);
 int    device_handle_free(device_handle dh);
 int    device_handle_free_all(clixon_handle h);
 device_handle device_handle_find(clixon_handle h, const char *name);
-device_handle  device_handle_each(clixon_handle h, device_handle dhprev);
+device_handle device_handle_each(clixon_handle h, device_handle dhprev);
 int    device_handle_connect(device_handle dh, clixon_client_type socktype, const char *dest,
                              int stricthostkey);
 int    device_handle_disconnect(device_handle dh);
@@ -97,6 +97,8 @@ char  *device_handle_schema_rev_get(device_handle dh);
 int    device_handle_schema_rev_set(device_handle dh, char  *schema_rev);
 char  *device_handle_logmsg_get(device_handle dh);
 int    device_handle_logmsg_set(device_handle dh, char *logmsg);
+char  *device_handle_domain_get(device_handle dh);
+int    device_handle_domain_set(device_handle dh, char *domain);
 cbuf  *device_handle_outmsg_get(device_handle dh, int nr);
 int    device_handle_outmsg_set(device_handle dh, int nr, cbuf *cb);